Don't attempt to drop files in non-existent paths
In MCollective 2.8 the packages no longer put files into core_libdir
(MCO-583, MCO-315), which makes these file resources fail as the
parent directory is never created but is assumed to be created by
the package install.

`refresh-mcollective-metadata` should never have been put into
core_libdir to begin with as that's intended to be space owned by
native OS packages.   It's doubly inappropriate as it's not a
loadable ruby file for an MCollective extension, it's a helper
executable for this module.

Here we just put it in $site_libdir, as that is a directory that
this module creates and populates.